requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
tensorflow-cpu==2.11.0
pandas==1.5.3
scikit-learn==1.2.2
indic-transliteration==2.3.43
numpy==1.24.4